% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlorichShp.R
\name{GlorichShp}
\alias{GlorichShp}
\title{Create a shapefile with all sampling locations}
\usage{
GlorichShp(
  data,
  coord_sys = list(c("NA1983", "4269"), c("WGS1984", "4326"), c("WGS 1984", "4326"),
    c("WGS84", "4326"), c("Gauss\\x96Kr\\xfcger, Zone 4, DHDN", "5678"),
    c("ERTS_UTM, Zone 33N, 7 digits", "25833"), c("British National Grid", "27700"),
    c("NatGrid NZ", "27200"), c("National Grid Australia", "5825")),
  convert = TRUE,
  final_CRS = "4326",
  export = TRUE,
  filename = NULL,
  filepath = NULL
)
}
\arguments{
\item{data}{The complete hydrochemistry dataset as produced by LoadHydroData
when 'location' is specified as true. Any subset of this dataset can also be
used as an input here - though the parameter 'coord_sys' should be updated
accordingly}

\item{coord_sys}{The is a list of vectors of length 2. Each vector in the list
contains: (1) The name of the coordinate reference system as it appears in
the 'CoordinateSystem' Column of the database and (2) the EPSG code for
said coordinate system. The default contains every unique system label in the
database as of 08/06/2022. Users are encouraged to check whether these are
correct before using the function}

\item{convert}{A logical vector indicating whether the user would like every
point to be converted to the same coordinate reference system}

\item{final_CRS}{The EPSG code for the coordinate reference system to which each
point should be converted. Must be enetered as a character string containing
only the digits of the code. Default is '4362' - the code for WGS84.}

\item{export}{A logical vector indicating whether the user would like the
created SpatialPointsDataFrames to be exported as .shp files. If TRUE (default)
either a complete shapefile, containing all data points in the specified
coordinate reference system or multiple shapefiles for each different system
will be exported depnding on whether 'convert' is TRUE or FALSE}

\item{filename}{A character string contaning the desired name of the file to export.
Only required if 'export' is TRUE. If 'convert' is TRUE a single character string
is required. If 'convert' is FALSE a list of filenames of the same length as the
list passed to 'coord_sys' is required}

\item{filepath}{The absolute filepath for the export location of the shapefiles.
Only required if 'export' is TRUE. Defaults to the current working directory.}
}
\value{
Either one large SpatialPointsDataFrame in the specified coordinate
  reference system or a list of SPDFs in the original coordinate reference systems
}
\description{
This function provides a number of options for generating shapefiles from the
GLORICH database. Individual shapefiles for each corrdinate system can be
created - or all coordinate systems can be converted to one specified system
and merged into a large shapefile containing all input points
}
\details{
.
}
\examples{
GlorichShp(data = hydrochemistry, convert = FALSE, export = FALSE)
  Returns a list of SPDFs of the same length as 'coord_sys' which have not been
  converted to a different coordinate reference system to the ones used in the database

GlorichShp(data = hydrochemistry, filename = 'GlorichShp')
  Return a SPDF containing every point in 'hydrochemistry'


}
