% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadHydroData.R
\name{LoadHydroData}
\alias{LoadHydroData}
\title{Import hydrochemistry and sampling location data}
\usage{
LoadHydroData(hydro_filepath, location = F, loc_filepath)
}
\arguments{
\item{hydro_filepath}{The file path to the hydrochemsitry.csv file (either an absolute path, or a relative path from the current wd)}

\item{location}{A boolean determining whether location data should be included, false by default}

\item{loc_filepath}{The file path to the sampling_location.csv file (either an absolute path, or a relative path from the current wd)}
}
\value{
A dataframe containing data from hydrochemistry.csv
}
\description{
This function imports the hydrochemstry.csv file from the GLORICH
database, and has the option to add location data from 'sampling_locations.csv'
to each sample.
}
\details{
The data columns from sampling_locations.csv that are added are:
'Country', 'State', 'Latitude', 'Longitude', 'CoordinateSystem'
.
}
\examples{
LoadHydroData(hydro_filepath = 'hydrochemistry.csv', location = T, loc_filepath = 'sampling_locations.csv')
LoadHydroData(hydro_filepath = 'hydrochemistry.csv')
}
