% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadHydroData.R
\name{LoadHydroData}
\alias{LoadHydroData}
\title{Import hydrochemistry and sampling location data}
\usage{
LoadHydroData(location)
}
\arguments{
\item{location}{A boolean determining whether location data should be included}
}
\value{
A dataframe containing data from hydrochemistry.csv
}
\description{
This function imports the hydrochemstry.csv file from the GLORICH
database, and has the option to add location data from 'sampling_locations.csv'
to each sample.
}
\details{
The data columns from sampling_locations.csv that are added are:
'Country', 'State', 'Latitude', 'Longitude', 'CoordinateSystem'
.
}
\examples{
LoadHydroData(location = F)
}
