% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadHydroData.R
\name{LoadHydroData}
\alias{LoadHydroData}
\title{Imports hydrochemsitry.csv as a data frame and optionally combines samples
with location data}
\usage{
LoadHydroData(location)
}
\arguments{
\item{location}{A boolean determining whether location data should be included}
}
\value{
A dataframe containing data from hydrochemistry.csv
}
\description{
This function converts imports the hydrochemstry.csv file from the GLORICH
database, and has the option to add location data from 'sampling_locations.csv'
to each sample
.
}
\examples{
LoadHydroData(location = F)
}
